+++ brdr/grb.py
@@ -7,7 +7,7 @@ from brdr.constants import (
     DOWNLOAD_LIMIT,
     DEFAULT_CRS,
     GRB_FEATURE_URL,
-    GRB_FISCAL_PARCELS_URL,
+    GRB_FISCAL_PARCELS_URL, MAX_REFERENCE_BUFFER,
 )
 from brdr.enums import GRBType, Evaluation
 from shapely import intersects, make_valid
@@ -15,8 +15,9 @@ from shapely.geometry import shape
 
 from brdr.geometry_utils import (
     features_by_geometric_operation,
-    create_donut, get_bbox
+    create_donut, get_bbox, buffer_pos
 )
+from brdr.loader import GeoJsonLoader
 from brdr.utils import get_collection, dict_series_by_keys, get_collection_by_partition, geojson_to_dicts
 
 log = logging.getLogger(__name__)
@@ -274,3 +275,36 @@ def check_equality(base_formula, actual_formula,threshold_area=5,threshold_perce
     if base_formula["full"] and base_formula["full"]:
         return True, Evaluation.EQUALITY_GEOM_3
     return False, Evaluation.NO_PREDICTION_5
+
+class GRBActualLoader(GeoJsonLoader):
+    def __init__(self, grb_type: GRBType,aligner, partition: int=1000):
+        self.aligner = aligner
+        self.grb_type = grb_type
+        self.part = partition
+        super().__init__()
+
+    def load_data(self):
+        if not self.aligner.dict_thematic:
+            raise ValueError("Thematic data not loaded")
+        geom_union = buffer_pos(self.aligner._get_thematic_union(), MAX_REFERENCE_BUFFER)
+        collection, id_property = get_collection_grb_actual(grb_type=self.grb_type, geometry=geom_union, partition=self.part,crs=self.aligner.CRS)
+        self.id_property = id_property
+        self.input = dict(collection)
+        self.aligner.logger.feedback_info(f"GRB downloaded: {self.grb_type}")
+        return super().load_data()
+
+class GRBFiscalParcelLoader(GeoJsonLoader):
+    def __init__(self, year: str, aligner,partition=1000):
+        self.aligner = aligner
+        self.year = year
+        self.part = partition
+        super().__init__(_input=None, id_property="CAPAKEY")
+
+    def load_data(self):
+        if not self.aligner.dict_thematic:
+            raise ValueError("Thematic data not loaded")
+        geom_union = buffer_pos(self.aligner._get_thematic_union(),MAX_REFERENCE_BUFFER)
+        collection = get_collection_grb_fiscal_parcels(year=self.year, geometry=geom_union, partition=self.part, crs=self.aligner.CRS)
+        self.input = dict(collection)
+        self.aligner.logger.feedback_info(f"Adpf downloaded for year: {self.year}")
+        return super().load_data()
