+++ examples/example_combined_borders_adp_gbg.py
@@ -1,6 +1,6 @@
 from brdr.enums import GRBType
-from brdr.grb import get_collection_grb_actual
-from brdr.loader import GRBActualLoader, GeoJsonFileLoader
+from brdr.grb import get_collection_grb_actual, GRBActualLoader
+from brdr.loader import GeoJsonFileLoader
 from brdr.aligner import Aligner
 from brdr.utils import polygonize_reference_data, geojson_to_dicts
 from examples import show_map, print_formula
